// src/common/permissions/permission.enum.ts
export enum Permission {
  ORDER_CANCEL_LATE = 'ORDER_CANCEL_LATE',
  ORDER_MARK_FREE = 'ORDER_MARK_FREE',
  ORDER_DISCOUNT = 'ORDER_DISCOUNT',
  ORDER_REFUND = 'ORDER_REFUND',
  REPORT_VIEW = 'REPORT_VIEW',
  INVENTORY_ADJUST = 'INVENTORY_ADJUST',
  PAYROLL_CONFIRM = 'PAYROLL_CONFIRM',
  PRINT_QUEUE_MANAGE = 'PRINT_QUEUE_MANAGE',
  INVOICE_VIEW = 'INVOICE_VIEW',
  INVOICE_PRINT_REQUEST = 'INVOICE_PRINT_REQUEST',
  CASHIER_SHIFT_OPEN = 'CASHIER_SHIFT_OPEN',
  CASHIER_SHIFT_CLOSE = 'CASHIER_SHIFT_CLOSE',
  CASHIER_SHIFT_VIEW_HISTORY = 'CASHIER_SHIFT_VIEW_HISTORY',
  STAFF_PERMISSION_MANAGE = 'STAFF_PERMISSION_MANAGE',
  MENU_MANAGE = 'MENU_MANAGE',
  TABLE_MANAGE = 'TABLE_MANAGE',
}